+++ src/commands/modules/deploy-stop-by-user.error.js
@@ -1,0 +1,5 @@
+export class DeployStopedByUser extends Error {
+  constructor(message = 'User stop the deployment phase.') {
+    super(message)
+  }
+}