+++ src/services/deploy/index.js
@@ -1,7 +1,0 @@
-const { DeployService } = require('./deploy.service')
-const { TargetFileAlreadyExist } = require('./target-file-already-exist.error')
-
-module.exports = exports = {
-  DeployService,
-  TargetFileAlreadyExist
-}