+++ src/shared/errors/bad-script-permission.error.js
@@ -1,9 +1,7 @@
-class BadScriptPermission extends Error {
+export class BadScriptPermission extends Error {
   constructor(scriptName, message = `Bad permission on "${scriptName}" script.`) {
     super(message)
 
     this.scriptName = scriptName
   }
 }
-
-module.exports = exports = { BadScriptPermission }