+++ src/shared/errors/folder-not-empty.error.js
@@ -1,7 +1,5 @@
-class FolderNotEmpty extends Error {
+export class FolderNotEmpty extends Error {
   constructor(message = 'Folder is not empty.') {
     super(message)
   }
 }
-
-module.exports = exports = { FolderNotEmpty }