+++ src/shared/errors/script-not-exist.error.js
@@ -1,9 +1,7 @@
-class ScriptNotExist extends Error {
+export class ScriptNotExist extends Error {
   constructor(scriptName, message = `Script "${scriptName}" doesn't exist.`) {
     super(message)
 
     this.scriptName = scriptName
   }
 }
-
-module.exports = exports = { ScriptNotExist }