+++ src/shared/errors/target-file-already-exist.error.js
@@ -1,9 +1,7 @@
-class TargetFileAlreadyExist extends Error {
+export class TargetFileAlreadyExist extends Error {
   constructor(filename, message = `Target file "${filename}" already exist.`) {
     super(message)
 
     this.filename = filename
   }
 }
-
-module.exports = exports = { TargetFileAlreadyExist }