+++ tests/services/exec/exec.service.js
@@ -3,8 +3,8 @@ import fs from 'fs'
 import path from 'path'
 import rimraf from 'rimraf'
 
-import { ExecService } from '../../src/services/exec.service'
-import { ScriptNotExist } from '../../src/services/exec'
+import { ExecService } from '../../../src/services/exec.service'
+import { ScriptNotExist } from '../../../src/services/exec'
 
 const tmpData = path.join(__dirname, './env-file')