+++ tests/services/file.service.js
@@ -3,7 +3,7 @@ import fs from 'fs'
 import path from 'path'
 import rimraf from 'rimraf'
 
-import { FileService } from '../../src/services/file.service'
+import { FileService } from '../../src'
 
 const tmpData = path.join(__dirname, './env-file')